refactor(foundation): untie tests from dcos-ui
