+++ foundation-ui/src/mount/__tests__/Mount-test.js
@@ -1,7 +1,3 @@
-jest.dontMock('../Mount');
-jest.dontMock('../MountService');
-jest.dontMock('../index');
-jest.dontMock('../../utils/ReactUtil');
 /* eslint-disable no-unused-vars */
 const React = require('react');
 /* eslint-enable no-unused-vars */
