+++ foundation-ui/src/mount/__tests__/MountService-test.js
@@ -1,4 +1,3 @@
-jest.dontMock('../MountService');
 /* eslint-disable no-unused-vars */
 const React = require('react');
 /* eslint-enable no-unused-vars */
