+++ foundation-ui/src/utils/__tests__/ReactUtil-test.js
@@ -1,4 +1,3 @@
-jest.dontMock('../ReactUtil');
 /* eslint-disable no-unused-vars */
 const React = require('react');
 /* eslint-enable no-unused-vars */
